# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Genrel Api",
        "description": "Apis for Goodreads Website!",
        "version": "1.0.0"
    },
    "servers": [{
        "url": "http://localhost:3000/"
    }],
    "security": [{
        "bearerAuth": []
    }],
    "paths": {
...
        "/genre": {
            "get": {
                "tags": [
                    "Genre"
                ],
                "summary": "Get All Genres",
                "responses": {
                    "200": {
                        "description": "all users"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/genre/add-genre": {
            "post": {
                "tags": [
                    "Genre"
                ],
                "summary": " Add Genres",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Genre Added"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "409": {
                        "description": "Genre All Ready Exist"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/genre/edit-genre/{id}": {
            "patch": {
                "tags": [
                    "Genre"
                ],
                "summary": "Edit a Genre.",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "genre id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Genre edited"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
        "/genre/delete/{id}": {
            "delete": {
                "tags": [
                    "Genre"
                ],
                "summary": "Delete a Genre.",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "genre id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Genre Delete"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
...
        "/book/genre": {
            "post": {
                "tags": [
                    "Book"
                ],
                "summary": "Get a Book by genre",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "Genre_id": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Get Books"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
...
        "/read/": {
            "get": {
                "tags": ["Read"],
                "summary": "Get All Reads",
                "responses": {
                    "200": {
                        "description": "Get All Reads"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/read/add-read": {
            "post": {
                "tags": ["Read"],
                "summary": "create Read",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "book_id": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Read Added"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "409": {
                        "description": "Read All Ready Exist"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/read/read-complete": {
            "patch": {
                "tags": ["Read"],
                "summary": "Change Read Status to complete so review can be given",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "number"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Read Completed"
                    },
                    "401": {
                        "description": "Auth Fail"
                    },
                    "409": {
                        "description": "Read Does not Exist"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/read/delete/{id}": {
            "delete": {
                "tags": [
                    "Read"
                ],
                "summary": "Delete a Read.",
                "parameters": [{
                    "name": "id",
                    "in": "path",
                    "description": "read id",
                    "required": true,
                    "style": "simple",
                    "explode": false,
                    "schema": {
                        "type": "number"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "read Delete"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    },
                    "401": {
                        "description": "Auth Fail"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}